refactored and fixed photo-comments route was very broken on mobile, and somewhat on desktop. deduplicated implementations of the same use cases by creating a photo component and using that to display photo and comments in photo-album as well as photo-comments
