+++ app/components/photo-albums/photo.hbs
@@ -1,0 +1,119 @@
+<div class='d-flex justify-content-center align-items-center'>
+  <img src={{@model.imageUrl}} class='photo-large' />
+</div>
+
+{{#if (can 'show individual users')}}
+  <p class='text-right text-muted'>
+    Foto geüpload door
+    <LinkTo @route='users.show' @model={{@model.uploader.id}}>
+      {{@model.uploader.fullName}}
+    </LinkTo>
+
+    {{#if @model.hasExif}}
+      <button
+        class='ml-1 btn btn-sm btn-outline-info'
+        {{action 'toggleShowExif'}}
+        type='button'
+      >
+        Toon Exif data
+      </button>
+    {{else}}
+      <button class='ml-1 btn btn-sm btn-outline-info disabled' type='button'>
+        Geen Exif data
+      </button>
+    {{/if}}
+  </p>
+
+  {{#if showExif}}
+    <div class='card card-body'>
+      <table class='table table-sm table-responsive text-nowrap'>
+        <tr>
+          <th>Make</th>
+          <th>Model</th>
+          <th>Original Datetime</th>
+          <th>Exposure</th>
+          <th>Aperture</th>
+          <th>ISO</th>
+          <th>Copyright</th>
+          <th>Lens</th>
+          <th>Zoom</th>
+        </tr>
+        <tr>
+          <td>{{@model.exifMake}}</td>
+          <td>{{@model.exifModel}}</td>
+          <td>{{moment-format
+              @model.exifDateTimeOriginal
+              'DD-MM-YYYY HH:mm'
+            }}</td>
+          <td>{{@model.exifExposureTime}}</td>
+          <td>
+            {{#if @model.exifApertureValue}}
+              f/{{@model.exifApertureValue}}
+            {{/if}}
+          </td>
+          <td>{{@model.exifIsoSpeedRatings}}</td>
+          <td>{{@model.exifCopyright}}</td>
+          <td>{{@model.exifLensModel}}</td>
+          <td>
+            {{#if @model.exifFocalLength}}
+              {{@model.exifApertureValue}}mm
+            {{/if}}
+          </td>
+        </tr>
+      </table>
+    </div>
+  {{/if}}
+{{/if}}
+
+{{#if (can 'show photo comments of photo' model)}}
+  <div class='card'>
+    <div class='card-header'>
+      <h5> Reacties </h5>
+    </div>
+    <div class='card-body'>
+      <div class='col'>
+        <div class='photo-comments'>
+          {{#each @model.comments as |comment|}}
+            <div class='row'>
+              {{user-comment comment data-test-photo-comment=comment.id}}
+            </div>
+          {{else}}
+            <span class='text-muted no-comment-text'>
+              Er zijn nog geen reacties
+            </span>
+          {{/each}}
+        </div>
+      </div>
+    </div>
+    {{#if (can 'create photo-comments')}}
+      <div class='card-footer'>
+        {{#if @model.photoAlbum.publiclyVisible}}
+          <div class='alert alert-info'>
+            Deze foto is ook te zien voor niet-leden, dat geldt ook voor de
+            reacties die er geplaatst worden. Het is natuurlijk heel fijn om te
+            laten zien hoe actief Alpha is, maar houd er wel rekening mee als je
+            wat plaatst.
+          </div>
+        {{/if}}
+        <div class='input-group new-photo-comment'>
+          <Input
+            @value={{content}}
+            class='form-control'
+            placeholder='Typ hier je reactie'
+            @minlength='1'
+            @maxlength='500'
+            @enter={{action 'submitPhotoComment'}}
+            @id='new-photo-comment-input'
+          />
+          <span class='input-group-btn'>
+            <button
+              class='btn btn-default'
+              {{action 'submitPhotoComment'}}
+              type='button'
+            >Reageren</button>
+          </span>
+        </div>
+      </div>
+    {{/if}}
+  </div>
+{{/if}}
