+++ app/templates/photo-comments/index.hbs
@@ -14,52 +14,7 @@
 <div class='row'>
   <div class='col-12'>
     {{#each model as |photo|}}
-      <div
-        class='card photo-and-comments-card mb-2 mb-md-5'
-        data-test-photo-with-comments-card
-      >
-        <div class='d-flex'>
-          <LinkTo
-            @route='photo-albums.photo-album.photos.show'
-            @models={{array photo.photoAlbum photo}}
-            class='col-12 col-md-8 p-0'
-          >
-            <img
-              class='image'
-              src={{photo.imageUrl}}
-              style='height: 400px; width: 100%;'
-            />
-          </LinkTo>
-          <div
-            class='col-12 col-md-4 photo-comments-and-input-holder p-0 d-none d-md-flex'
-          >
-            <div class='photo-comment-scroll-container'>
-              <div class='photo-comments'>
-                {{#each photo.comments as |comment|}}
-                  {{user-comment comment data-test-user-comment=true}}
-                {{/each}}
-              </div>
-            </div>
-            <div class='input-group new-photo-comment bg-light p-3'>
-              <Input
-                @value={{mut (get commentContent (get photo 'id'))}}
-                class='form-control photo-comment-input'
-                placeholder='Typ hier je reactie'
-                @minlength='1'
-                @maxlength='500'
-                @enter={{action 'submitPhotoComment' photo}}
-              />
-              <span class='input-group-btn'>
-                <button
-                  class='btn btn-default submit-photo-comment-button'
-                  {{action 'submitPhotoComment' photo}}
-                  type='button'
-                >Reageren</button>
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
+      <PhotoAlbums::Photo @model={{photo}} />
     {{/each}}
   </div>
 </div>
